isolate top-level test modules
